Break watch.py into watch/ package with logical modules
watch/
  __init__.py    — public API (ActorWatchApp, run_watch, main)
  __main__.py    — python -m actor.watch entry point
  app.py         — main Textual App, compose, actions, tab management
  tree.py        — ActorTree widget
  themes.py      — claude-dark and claude-light theme definitions
  helpers.py     — group_by_parent, read_log_entries, compute_diff
  patches.py     — ANSIToTruecolor monkey-patch for transparency
