+++ src/actor/watch/__init__.py
@@ -1,0 +1,5 @@
+"""actor watch — real-time dashboard for actor.sh."""
+
+from .app import ActorWatchApp, run_watch, main
+
+__all__ = ["ActorWatchApp", "run_watch", "main"]
